html/semantics/embedded-content-0/media-elements/loading-the-media-resource/022.html

<!doctype html><base href=http://w3c-test.org>
<title>invoking resource selection by inserting document fragment into a document</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
var v;
var t = async_test(function(t) {
  v = document.createElement('video');
  var fragment = document.createDocumentFragment();
  fragment.appendChild(v);
  assert_equals(v.networkState, 0, 'networkState after fragment.appendChild(v)');
  document.body.appendChild(fragment);
  assert_equals(v.networkState, 3, 'networkState after document.body.appendChild(fragment)');
});
</script>
<script>
t.step(function() {
  assert_equals(v.networkState, 0, 'networkState in separate script');
  t.done();
});
</script>
